backend/requirements.txt
fastapi
uvicorn[standard]
crewai
crewai-tools
google-genai
arxiv
pydantic
python-dotenv
httpx

